+++ CHANGELOG.md
@@ -1,8 +1,22 @@
-# v0.5.0
+# v0.6.0
 
 ## New Features
 
 * [#5](https://github.com/harrisont/fastbuild-vscode/issues/5) Support "go to definition" and "find references" for target names.
+
+# v0.5.1
+
+## Bug Fixes
+
+* `#if exists(...)` now correctly evaluates if the environment variable exists instead of always being false (#23).
+
+# v0.5.0
+
+## New Features
+
+* Support "go to definition" and "find references" for `#define`s:
+    * [`#define` / `#undef` FASTBuild docs](https://www.fastbuild.org/docs/syntaxguide.html#define)
+    * [`#if` / `#else` / `#endif` FASTBuild docs](https://www.fastbuild.org/docs/syntaxguide.html#if)
 
 # v0.4.0